refactor: add interface file suffix for groups
