+++ plugin-server/src/ingestion/ingestion-consumer.ts
@@ -28,7 +28,7 @@ import { PromiseScheduler } from '../utils/promise-scheduler'
 import { retryIfRetriable } from '../utils/retries'
 import { EventPipelineResult, EventPipelineRunner } from '../worker/ingestion/event-pipeline/runner'
 import { BatchWritingGroupStore } from '../worker/ingestion/groups/batch-writing-group-store'
-import { GroupStoreForBatch } from '../worker/ingestion/groups/group-store-for-batch'
+import { GroupStoreForBatch } from '../worker/ingestion/groups/group-store-for-batch.interface'
 import { ClickhouseGroupRepository } from '../worker/ingestion/groups/repositories/clickhouse-group-repository'
 import { PostgresGroupRepository } from '../worker/ingestion/groups/repositories/postgres-group-repository'
 import { BatchWritingPersonsStore } from '../worker/ingestion/persons/batch-writing-person-store'
