+++ plugin-server/src/worker/ingestion/event-pipeline/runner.ts
@@ -8,7 +8,7 @@ import { timeoutGuard } from '../../../utils/db/utils'
 import { normalizeProcessPerson } from '../../../utils/event'
 import { logger } from '../../../utils/logger'
 import { captureException } from '../../../utils/posthog'
-import { GroupStoreForBatch } from '../groups/group-store-for-batch'
+import { GroupStoreForBatch } from '../groups/group-store-for-batch.interface'
 import { PersonsStoreForBatch } from '../persons/persons-store-for-batch'
 import { EventsProcessor } from '../process-event'
 import { captureIngestionWarning, generateEventDeadLetterQueueMessage } from '../utils'
